Change of usage: the ParticleLCCurrent is supposed to be used by other current generators and does not work in a "stand-alone" way. For point-charge distributions there is now PointChargeLCCurrent which works as a wrapper for ParticleLCCurrent.
