+++ pixi/src/main/java/org/openpixi/pixi/ui/util/yaml/currentgenerators/YamlPointChargeLCCurrent.java
@@ -1,14 +1,11 @@
 package org.openpixi.pixi.ui.util.yaml.currentgenerators;
 
 import org.openpixi.pixi.physics.Settings;
-import org.openpixi.pixi.physics.fields.currentgenerators.ParticleLCCurrent;
+import org.openpixi.pixi.physics.fields.currentgenerators.PointChargeLCCurrent;
 
 import java.util.ArrayList;
 
-/**
- * Created by dmueller on 9/3/15.
- */
-public class YamlParticleLCCurrent {
+public class YamlPointChargeLCCurrent {
 
 	/**
 	 * Direction of the current pulse (0 to d)
@@ -31,6 +28,16 @@ public class YamlParticleLCCurrent {
 	public Double longitudinalWidth;
 
 	/**
+	 * Option whether to use the monopole removal method.
+	 */
+	public Boolean useMonopoleRemoval = false;
+
+	/**
+	 * Option whether to use the dipole removal method.
+	 */
+	public Boolean useDipoleRemoval = false;
+
+	/**
 	 * List of charges on the pulse plane
 	 */
 	public ArrayList<YamlPointCharge> charges = new ArrayList<YamlPointCharge>();
@@ -43,12 +50,12 @@ public class YamlParticleLCCurrent {
 	 */
 	public boolean checkConsistency(Settings settings) {
 		if (direction >= settings.getNumberOfDimensions()) {
-			System.out.println("NewLCCurrent: direction index exceeds the dimensions of the system.");
+			System.out.println("PointChargeLCCurrent: direction index exceeds the dimensions of the system.");
 			return false;
 		}
 
 		if(Math.abs(orientation) != 1) {
-			System.out.println("NewLCCurrent: orientation must be either -1 or 1.");
+			System.out.println("PointChargeLCCurrent: orientation must be either -1 or 1.");
 			return false;
 		}
 
@@ -58,12 +65,12 @@ public class YamlParticleLCCurrent {
 		for(YamlPointCharge c : charges) {
 			// Check color vectors
 			if (c.amplitudeColorDirection.size() != numberOfComponents) {
-				System.out.println("NewLCCurrent: aColor vector does not have the right dimensions.");
+				System.out.println("PointChargeLCCurrent: aColor vector does not have the right dimensions.");
 				return false;
 			}
 			// Check location vectors
 			if(c.location.size() != effDim) {
-				System.out.println("NewLCCurrent: location vector does not have the right dimensions.");
+				System.out.println("PointChargeLCCurrent: location vector does not have the right dimensions.");
 				return false;
 			}
 
@@ -74,8 +81,8 @@ public class YamlParticleLCCurrent {
 	}
 
 
-	public ParticleLCCurrent getCurrentGenerator() {
-		ParticleLCCurrent generator = new ParticleLCCurrent(direction, orientation, location, longitudinalWidth);
+	public PointChargeLCCurrent getCurrentGenerator() {
+		PointChargeLCCurrent generator = new PointChargeLCCurrent(direction, orientation, location, longitudinalWidth, useMonopoleRemoval, useDipoleRemoval);
 
 		for(YamlPointCharge c: charges) {
 			double[] chargeLocation = new double[c.location.size()];
